# frontend/src/lib/api/openapi.json
{"openapi":"3.1.0","info":{"title":"Tahir's List API","description":"HTTP API for Tahir's List. Contract lives at openapi/v1.json (OpenAPI 3.1).","version":"v1"},"servers":[{"url":"http://127.0.0.1:8080","description":"Generated server url"}],"paths":{"/v1/listings":{"post":{"tags":["listing-controller"],"summary":"Add a restaurant listing","description":"Creates a new, always-unverified restaurant listing owned by the authenticated account. Does not geocode; the client supplies coordinates.","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateListingRequest"}}},"required":true},"responses":{"201":{"description":"Listing created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ListingResponse"}}}},"400":{"description":"Invalid input","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Authentication required","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Owning account not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/v1/auth/signup":{"post":{"tags":["signup-controller"],"summary":"Create an account","description":"Validates the email/password, hashes the password with Argon2id, persists the account with the default role, and returns the created account.","operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"409":{"description":"Email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Password is too weak","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Invalid input","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"201":{"description":"Account created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SignupResponse"}}}}}}},"/v1/auth/refresh":{"post":{"tags":["refresh-controller"],"summary":"Refresh the session","description":"Rotates the refresh token presented in the HttpOnly cookie: revokes it and issues a fresh access token in the body plus a new refresh cookie.","operationId":"refresh","parameters":[{"name":"refresh_token","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"401":{"description":"Invalid, expired, or already-used refresh token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Missing refresh cookie","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"Rotated; new access token + refresh cookie issued","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/v1/auth/logout":{"post":{"tags":["logout-controller"],"summary":"Log out","description":"Revokes the refresh token presented in the HttpOnly cookie so it can no longer be used to obtain a fresh access token, and clears the cookie. Idempotent: always succeeds regardless of whether the token was ever valid.","operationId":"logout","parameters":[{"name":"refresh_token","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"204":{"description":"Refresh token revoked; cookie cleared"}}}},"/v1/auth/login":{"post":{"tags":["login-controller"],"summary":"Log in","description":"Verifies the email/password against the stored Argon2id hash and returns a short-lived access token (RS256 JWT with the account's RBAC role) in the body; the rotating refresh token is set as an HttpOnly; Secure; SameSite=Lax cookie scoped to the auth routes and is never returned in JSON.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"401":{"description":"Invalid email or password (generic, no user enumeration)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Invalid input","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"Authenticated; token pair issued","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/v1/me":{"get":{"tags":["me-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/v1/health":{"get":{"tags":["health-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}}},"components":{"schemas":{"ErrorResponse":{"properties":{"code":{"type":"string"},"message":{"type":"string"}},"required":["code"]},"CreateListingRequest":{"properties":{"name":{"type":"string"},"address":{"type":"string"},"lat":{"type":"number","format":"double"},"lng":{"type":"number","format":"double"},"cuisine":{"type":"string"},"cuttingMethod":{"type":"string","enum":["HAND_CUT","MACHINE_CUT","UNSPECIFIED"]}},"required":["address","cuisine","cuttingMethod","lat","lng","name"]},"ListingResponse":{"properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"address":{"type":"string"},"lat":{"type":"number","format":"double"},"lng":{"type":"number","format":"double"},"cuisine":{"type":"string"},"cuttingMethod":{"type":"string"},"ownerId":{"type":"string","format":"uuid"},"verificationStatus":{"type":"string","enum":["UNVERIFIED","VERIFIED"]},"createdAt":{"type":"string","format":"date-time"}},"required":["address","createdAt","cuisine","cuttingMethod","id","lat","lng","name","ownerId","verificationStatus"]},"SignupRequest":{"properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"SignupResponse":{"properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"role":{"type":"string"}},"required":["email","id","role"]},"AuthResponse":{"properties":{"accessToken":{"type":"string"},"tokenType":{"type":"string"},"expiresIn":{"type":"integer","format":"int64"},"accountId":{"type":"string","format":"uuid"},"role":{"type":"string"}},"required":["accessToken","accountId","expiresIn","role","tokenType"]},"LoginRequest":{"properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]}}}}